GameRunner - move memory related operations to new class 'Memory.java'
